Authenciation, session, and cookies implementation

+++ main/views.py
@@ -3,12 +3,19 @@ from django.contrib import messages
 from django.core import serializers
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404, redirect, render
+from django.contrib.auth import login, logout
+from django.contrib.auth.forms import AuthenticationForm, UserCreationForm
+from django.shortcuts import redirect, render
+from django.contrib.auth.decorators import login_required  
+from django.core.exceptions import PermissionDenied        
 
 from main.models import Experience
 from main.models import Project
 from main.forms import ProjectForm, ExperienceForm
+import datetime
 
 def show_main(request):
+    last_login = request.COOKIES.get('last_login', 'No active login session / Cookie not found')
     context = {
         "name": "Kevin Nicholas Santoso",
         "npm": "2506637041",
@@ -17,6 +24,7 @@ def show_main(request):
             "CS student at Universital Indonesia International Class. "
             "Part time artist drawing animals in their free time."
         ),
+        "last_login": last_login
     }
     return render(request, "index.html", context)
 
@@ -91,8 +99,11 @@ def project_list(request):
     }
     return render(request, "projects.html", context)
 
+@login_required(login_url="/login/")
 def create_project(request):
     form = ProjectForm(request.POST or None)
+    if not request.user.is_superuser:
+        raise PermissionDenied
 
     if request.method == "POST" and form.is_valid():
         form.save()
@@ -100,7 +111,7 @@ def create_project(request):
         return redirect("main:project_list")
 
     context = {
-        "name": "Burhan",
+        "name": "Kevin Nicholas Santoso",
         "form": form,
     }
     return render(request, "projects_form.html", context)
@@ -115,12 +126,63 @@ def get_projects_json(request):
     projects_json = serializers.serialize("json", projects)
     return HttpResponse(projects_json, content_type="application/json")
 
+@login_required(login_url="/login/")
 def delete_project(request, project_id):
+    if not request.user.is_superuser:
+        raise PermissionDenied
     project = get_object_or_404(Project, pk=project_id)
 
     if request.method == "POST":
         project.delete()
         messages.success(request, "Project berhasil dihapus!")
         return redirect("main:project_list")
+
+    return redirect("main:project_list")
+
+def register(request):
+    form = UserCreationForm(request.POST or None)
+
+    if request.method == "POST" and form.is_valid():
+        form.save()
+        messages.success(request, "Account created successfully. Please log in.")
+        return redirect("main:login")
+
+    context = {
+        "name": "Kevin Nicholas Santoso",
+        "form": form,
+    }
+    return render(request, "register.html", context)
+
+def login_user(request):
+    form = AuthenticationForm(request, data=request.POST or None)
+
+    if request.method == "POST" and form.is_valid():
+        user = form.get_user()
+        login(request, user)
+        response = redirect("main:show_main")
+        response.set_cookie('last_login', datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S'))
+        return response
+
+    context = {
+        "name": "Kevin Nicholas Santoso",
+        "form": form,
+    }
+    return render(request, "login.html", context)
+
+def logout_user(request):
+    logout(request)
+    response = redirect("main:show_main")
+    response.delete_cookie('last_login')
+    return response
+
+@login_required(login_url="/login/")
+def toggle_star(request, project_id):
+    project = get_object_or_404(Project, pk=project_id)
+
+    if request.method == "POST":
+        if request.user in project.starred_by.all():
+            project.starred_by.remove(request.user)
+        else:
+            project.starred_by.add(request.user)
 
     return redirect("main:project_list")
